requirements.txt
django==1.10.8
allauth
certifi (2018.1.18)
chardet (3.0.4)
Django (1.11.10)
django-allauth (0.35.0)
django-scheduler (0.8.8)
djangorestframework (3.7.7)
icalendar (4.0.1)
idna (2.6)
oauthlib (2.0.6)
pip (9.0.1)
pkg-resources (0.0.0)
PyJWT (1.6.1)
python-dateutil (2.6.1)
python-openid (2.2.5)
pytz (2018.3)
requests (2.18.4)
requests-oauthlib (0.8.0)
setuptools (38.5.1)
simplejson (3.13.2)
six (1.11.0)
social-auth-app-django (2.1.0)
social-auth-core (1.7.0)
stripe (1.79.0)
urllib3 (1.22)
wheel (0.30.0)
